+++ compiler/lib/generate.ml
@@ -152,7 +152,8 @@ module Share = struct
     match c with
     | String s -> add_code_string s t
     | NativeString s -> add_code_native_string s t
-    | Tuple (_, args, _) -> Array.fold_left args ~init:t ~f:(fun t c -> get_constant c t)
+    | Tuple (_, args, _, _) ->
+        Array.fold_left args ~init:t ~f:(fun t c -> get_constant c t)
     | _ -> t
 
   let add_args args t =
@@ -525,9 +526,10 @@ let rec constant_rec ~ctx x level instrs =
       | Utf (Utf8 x) -> Share.get_utf_string str_js_utf8 x ctx.Ctx.share, instrs)
   | Float f -> float_const f, instrs
   | Float32 f -> float_const f, instrs
-  | Float_array a ->
+  | Float_array (a, desc) ->
       ( Mlvalue.Array.make
           ~tag:Obj.double_array_tag
+          ~desc
           ~args:(Array.to_list (Array.map a ~f:(fun x -> J.Element (float_const x))))
       , instrs )
   | Int64 i ->
@@ -538,10 +540,10 @@ let rec constant_rec ~ctx x level instrs =
       and mi = int (Int64.to_int (Int64.logand (Int64.shift_right i 24) 0xffffffL))
       and hi = int (Int64.to_int (Int64.logand (Int64.shift_right i 48) 0xffffL)) in
       J.call p [ lo; mi; hi ] J.N, instrs
-  | Tuple (tag, a, _) -> (
+  | Tuple (tag, a, _, desc) -> (
       let constant_max_depth = Config.Param.constant_max_depth () in
       let rec detect_list n acc = function
-        | Tuple (0, [| x; l |], _) -> detect_list (succ n) (x :: acc) l
+        | Tuple (0, [| x; l |], _, _) -> detect_list (succ n) (x :: acc) l
         | Int maybe_zero when Targetint.is_zero maybe_zero ->
             if n > constant_max_depth then Some acc else None
         | _ -> None
@@ -579,7 +581,7 @@ let rec constant_rec ~ctx x level instrs =
                   | _ -> J.Element js :: acc, instrs)
             else List.map ~f:(fun x -> J.Element x) (List.rev l), instrs
           in
-          Mlvalue.Block.make ~tag ~args:l, instrs)
+          Mlvalue.Block.make ~tag ~desc ~args:l, instrs)
   | Int i -> targetint i, instrs
   | Int32 i | NativeInt i -> targetint (Targetint.of_int32_exn i), instrs
   | Null_ -> s_var "null", instrs
@@ -1503,7 +1505,7 @@ let rec translate_expr ctx loc x e level : (_ * J.statement_list) Expr_builder.t
       let* args = list_map (access ~ctx) args in
       let* f = access ~ctx f in
       return (apply_fun ctx f args exact trampolined in_cps loc, [])
-  | Block (tag, a, array_or_not, _mut) ->
+  | Block (tag, a, array_or_not, _mut, desc) ->
       let* contents =
         list_map
           (fun x ->
@@ -1521,8 +1523,8 @@ let rec translate_expr ctx loc x e level : (_ * J.statement_list) Expr_builder.t
       in
       let x =
         match array_or_not with
-        | Array -> Mlvalue.Array.make ~tag ~args:contents
-        | NotArray | Unknown -> Mlvalue.Block.make ~tag ~args:contents
+        | Array -> Mlvalue.Array.make ~tag ~desc ~args:contents
+        | NotArray | Unknown -> Mlvalue.Block.make ~tag ~desc ~args:contents
       in
       return (x, [])
   | Field (x, n, _) ->