+++ compiler/lib/mlvalue.mli
@@ -18,15 +18,26 @@
  *)
 
 module Block : sig
-  val make : tag:int -> args:Javascript.element list -> Javascript.expression
+  val make :
+       tag:int
+    -> desc:Code.block_desc
+    -> args:Javascript.element list
+    -> Javascript.expression
+  (** Introcaml: with introspection enabled, [desc] is stored in the header
+      word, above the tag. *)
 
   val tag : Javascript.expression -> Javascript.expression
+  (** The tag of a block, without the descriptor bits. *)
 
   val field : Javascript.expression -> int -> Javascript.expression
 end
 
 module Array : sig
-  val make : tag:int -> args:Javascript.element list -> Javascript.expression
+  val make :
+       tag:int
+    -> desc:Code.block_desc
+    -> args:Javascript.element list
+    -> Javascript.expression
 
   val length : Javascript.expression -> Javascript.expression